pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iris"
version = "0.1.0"
description = "Web Automation Engine — Playwright-based page fetching for the Cortex ecosystem"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "playwright>=1.49.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "httpx>=0.27.0",
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "PyJWT>=2.9.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "fakeredis>=2.26.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "types-redis>=4.6.0",
    "types-beautifulsoup4>=4.12.0",
    "lxml-stubs>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/iris"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "A", "SIM",
    "RUF", "PTH", "ASYNC",
]

[tool.ruff.lint.per-file-ignores]
"src/iris/sentinel_sdk/client.py" = ["ASYNC109"]

[tool.ruff.lint.isort]
known-first-party = ["iris"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "trafilatura.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["iris"]

[tool.coverage.report]
fail_under = 80
show_missing = true
